scraping/kym-scrapper/requirements.txt
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
coverage==7.3.2
cryptography==41.0.5
cssselect==1.2.0
distlib==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.3
filelock==3.13.1
flake8==6.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
mccabe==0.7.0
mock==5.1.0
packaging==23.2
parsel==1.8.1
platformdirs==3.11.0
pluggy==1.3.0
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
PyDispatcher==2.0.7
pyflakes==3.1.0
pymongo==4.6.0
pyOpenSSL==23.3.0
pyproject-api==1.6.1
pytest==7.4.3
pytest-cov==4.1.0
queuelib==1.6.2
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-redis==0.7.3
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
tldextract==5.1.0
tomli==2.0.1
tox==4.11.3
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.6
w3lib==2.1.2
zope.interface==6.1
